Replacing code that's now in the new Tinder project with a jar generated by the Tinder project (OF-1, TINDER-5).
git-svn-id: http://svn.igniterealtime.org/svn/repos/openfire/trunk@11030 b35dd754-fafc-0310-a699-88a17e54d16e

+++ src/java/org/jivesoftware/openfire/component/InternalComponentManager.java
@@ -263,7 +263,7 @@ public class InternalComponentManager extends BasicModule implements ComponentMa
         }
     }
 
-    public IQ query(Component component, IQ packet, int timeout) throws ComponentException {
+    public IQ query(Component component, IQ packet, long timeout) throws ComponentException {
         final LinkedBlockingQueue<IQ> answer = new LinkedBlockingQueue<IQ>(8);
         XMPPServer.getInstance().getIQRouter().addIQResultListener(packet.getID(), new IQResultListener() {
             public void receivedAnswer(IQ packet) {
